# api/openapi.yaml
openapi: 3.1.0
info:
  title: KakEnBetaal API
  version: 0.1.0
  description: HTTP contract. Money is integer cents. Keep in sync with code; types are generated with `pnpm gen:api`.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearer: {type: http, scheme: bearer, bearerFormat: JWT}
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties: {code: {type: string}, message: {type: string}, details: {type: object}}
...
    Business:
      type: object
      properties:
        name: {type: string}
        tradingName: {type: [string, 'null']}
        vatRegistered: {type: boolean}
        vatNumber: {type: [string, 'null'], pattern: '^4[0-9]{9}$'}
        address: {type: [string, 'null']}
    ReportSummary:
      type: object
      required: [from, to, count, grossCents, tipCents, feeCents, refundCents, netCents, reconciled, ledger, byDay, byParty, byService]
      properties:
        from: {type: string, format: date}
        to: {type: string, format: date}
        count: {type: integer}
        grossCents: {type: integer}
        baseCents: {type: integer}
        tipCents: {type: integer}
        feeCents: {type: integer}
        refundCount: {type: integer}
        refundCents: {type: integer}
        netCents: {type: integer}
        reconciled: {type: boolean, description: The same totals summed from the ledger agree}
        ledger: {type: object, properties: {creditsCents: {type: integer}, feeCents: {type: integer}, refundCents: {type: integer}}}
        byDay: {type: array, items: {type: object, properties: {date: {type: string, format: date}, count: {type: integer}, grossCents: {type: integer}, tipCents: {type: integer}}}}
        byParty: {type: array, items: {type: object, properties: {partyKind: {type: string}, userId: {type: [string, 'null']}, name: {type: [string, 'null']}, salesCents: {type: integer}, tipCents: {type: integer}, feeCents: {type: integer}, refundCents: {type: integer}, netCents: {type: integer}}}}
        byService: {type: array, items: {type: object, properties: {serviceId: {type: [string, 'null']}, name: {type: string}, count: {type: integer}, amountCents: {type: integer}}}}
...
  responses:
    Err: {description: Error, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
security: []
paths:
...
  /v1/merchant/services:
    get: {tags: [merchant], security: [{bearer: []}], responses: {'200': {description: list}}}
    post: {tags: [merchant], security: [{bearer: []}], summary: Add a service (manager), requestBody: {required: true, content: {application/json: {schema: {type: object, required: [name, priceCents], properties: {name: {type: string}, priceCents: {type: integer, minimum: 0}}}}}}, responses: {'201': {description: created}}}
  /v1/merchant/staff:
    get: {tags: [merchant], security: [{bearer: []}], responses: {'200': {description: list}}}
    post: {tags: [merchant], security: [{bearer: []}], summary: Add staff by WhatsApp number (manager). They sign in with a WhatsApp code; only an owner adds an owner., requestBody: {required: true, content: {application/json: {schema: {type: object, required: [name, msisdn], properties: {name: {type: string}, msisdn: {type: string}, role: {type: string, enum: [staff, manager, owner]}}}}}}, responses: {'201': {description: added}, '403': {$ref: '#/components/responses/Err'}, '409': {$ref: '#/components/responses/Err'}, '422': {$ref: '#/components/responses/Err'}}}
...
  /v1/merchant/reports/summary: {get: {tags: [merchant], security: [{bearer: []}], summary: Totals for a period with the ledger cross-check. Staff see their own bills and earnings., parameters: [{name: from, in: query, schema: {type: string, format: date}, description: 'SAST date, default today'}, {name: to, in: query, schema: {type: string, format: date}, description: 'SAST date, default from; at most 366 days'}], responses: {'200': {description: totals, content: {application/json: {schema: {$ref: '#/components/schemas/ReportSummary'}}}}, '422': {$ref: '#/components/responses/Err'}}}}
  /v1/merchant/reports/export.csv: {get: {tags: [merchant], security: [{bearer: []}], summary: One row per payment (manager). Amounts in rand with a dot; formula cells neutralised., parameters: [{name: from, in: query, schema: {type: string, format: date}, description: 'SAST date, default today'}, {name: to, in: query, schema: {type: string, format: date}, description: 'SAST date, default from; at most 366 days'}], responses: {'200': {description: csv, content: {text/csv: {schema: {type: string}}}}, '403': {$ref: '#/components/responses/Err'}}}}
  /v1/merchant/business:
    get: {tags: [merchant], security: [{bearer: []}], summary: Business details for slips and tax invoices (manager), responses: {'200': {description: details, content: {application/json: {schema: {$ref: '#/components/schemas/Business'}}}}}}
    patch: {tags: [merchant], security: [{bearer: []}], summary: Change business details. VAT registration needs a VAT number (10 digits, starts with 4) and an address., requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Business'}}}}, responses: {'200': {description: saved}, '422': {$ref: '#/components/responses/Err'}}}
  /v1/merchant/services/all: {get: {tags: [merchant], security: [{bearer: []}], summary: All services including hidden ones (manager), responses: {'200': {description: list}}}}
  /v1/merchant/services/{id}:
    patch: {tags: [merchant], security: [{bearer: []}], summary: Rename, reprice or hide a service (manager). Existing bills keep their price., parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}], requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {name: {type: string}, priceCents: {type: integer, minimum: 0}, active: {type: boolean}}}}}}, responses: {'200': {description: saved}, '404': {$ref: '#/components/responses/Err'}}}
  /v1/merchant/staff/{id}:
    patch: {tags: [merchant], security: [{bearer: []}], summary: Rename, change role or deactivate (manager; owners only by an owner; not yourself), parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}], requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {name: {type: string}, role: {type: string, enum: [staff, manager, owner]}, active: {type: boolean}}}}}}, responses: {'200': {description: saved}, '403': {$ref: '#/components/responses/Err'}, '404': {$ref: '#/components/responses/Err'}, '409': {$ref: '#/components/responses/Err'}}}
  /v1/merchant/payments/{id}/receipt:
    post: {tags: [merchant], security: [{bearer: []}], summary: Revoke the receipt link, or reissue it under a new token (manager). The old link stops working either way., parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}], requestBody: {required: true, content: {application/json: {schema: {type: object, required: [action], properties: {action: {type: string, enum: [revoke, reissue]}}}}}}, responses: {'200': {description: done, content: {application/json: {schema: {type: object, properties: {receiptUrl: {type: [string, 'null']}}}}}}, '404': {$ref: '#/components/responses/Err'}}}
...